dendrogram: Move implementation of DendrogramWidget
